+++ base-hack/asm/objects.asm
@@ -7,7 +7,6 @@
 .importobj "obj/src_fixes_guardCatch.o"
 .importobj "obj/src_fixes_helm.o"
 .importobj "obj/src_fixes_level_modifiers.o"
-.importobj "obj/src_fixes_overlay_changes.o"
 .importobj "obj/src_fixes_parent.o"
 .importobj "obj/src_fixes_quality_fixes.o"
 .importobj "obj/src_fixes_spiders.o"
@@ -64,6 +63,14 @@
 .importobj "obj/src_misc_tns_number.o"
 .importobj "obj/src_misc_warp_to_isles.o"
 .importobj "obj/src_misc_win_condition.o"
+.importobj "obj/src_overlays_arcade.o"
+.importobj "obj/src_overlays_bonus.o"
+.importobj "obj/src_overlays_boss.o"
+.importobj "obj/src_overlays_changes.o"
+.importobj "obj/src_overlays_critter.o"
+.importobj "obj/src_overlays_jetpac.o"
+.importobj "obj/src_overlays_menu.o"
+.importobj "obj/src_overlays_race.o"
 .importobj "obj/src_pause_carousel.o"
 .importobj "obj/src_pause_hints.o"
 .importobj "obj/src_pause_lib.o"
