+++ base-hack/include/global.h
@@ -419,6 +419,14 @@ extern void setKrushaAmmoColor(void);
 
 extern void loadWidescreen(overlays loaded_overlay);
 extern void initFilename(void);
+extern void* getPointerFile(int table, int file);
+extern void overlay_mod_menu(void);
+extern void overlay_mod_critter(void);
+extern void overlay_mod_boss(void);
+extern void overlay_mod_bonus(void);
+extern void initArcade(void);
+extern void initJetpac(void);
+extern void overlay_mod_race(void);
 
 extern void handleGrabbingLock(void* player, int player_index, int allow_vines);
 extern void handleLedgeLock(void);
