+++ base-hack/include/vars.h
@@ -155,3 +155,5 @@
 #define FUNGI_NIGHT_CHECK 0x100000
 #define ENABLE_CLIMBING_FLAG 0
 #define DISABLE_TRAINING_PRECHECKS 1
+
+#define ARCADE_IMAGE_COUNT 22
